requirements.txt
streamlit
openai
google-cloud-bigquery
python-dotenv
pandas
db-dtypes